Finished Task E: Randomized Rooms
Created tiles that are either blank, full blocked walls, or random blocks. The random blocks use a random boolean 4 times per tile to determine which blocks stay (and some will even stay as a tower). This forms a maze-like level when then blocks are in a grid. The blank floor tile can be used to ensure that there is always a path through the level, but this may not be necessary if a "reset level" button is added. It won't be used for now because in multiple tests the blocks were sparse enough to navigate in any direction.