Started Task H: Vehicle Menu
Created UI piece, added switch for the name of the exit button on the close popup ("next" or "close" based on the current level name). Used the same piece of code to open the UI piece for the vehicle menu. The vehicle selection, player stat effect (+saving) and the debt effect still need to be added